Implementation of Heat disatnce as a wrapper
