+++ geomfum/_registry.py
@@ -268,6 +268,10 @@ class MeshPlotterRegistry(WhichRegistry):
     MAP = {}
 
 
+class HeatDistanceMetricRegistry(WhichRegistry):
+    MAP = {}
+
+
 def _create_register_funcs(module):
     """Create ``register`` functions for each class registry in this module.
