+++ geomfum/metric/mesh.py
@@ -1,10 +1,11 @@
 """Module that metrics for calcualte distances on a mesh."""
+
 import abc
 
 import networkx as nx
 import numpy as np
-import potpourri3d as pp3d
 
+from geomfum._registry import HeatDistanceMetricRegistry, WhichRegistryMixins
 from geomfum.numerics.graph import single_source_partial_dijkstra_path_length
 
 
@@ -347,81 +348,15 @@ class KClosestGraphShortestPathMetric(_NxDijkstraMixins, FinitePointSetMetric):
         return np.array(list(dist_dict.values())), np.array(list(dist_dict.keys()))
 
 
-
-class HeatDistanceMetric(_SingleDispatchMixins,FinitePointSetMetric):
+class HeatDistanceMetric(
+    _SingleDispatchMixins, FinitePointSetMetric, WhichRegistryMixins
+):
     """Heat distance metric between vertices of a mesh.
-
-    Parameters
-    ----------
-    shape : Shape
-        Shape.
 
     References
     ----------
-    "The Heat Method for Distance Computation, Communications of the ACM (2017), 
+    "The Heat Method for Distance Computation, Communications of the ACM (2017),
     Keenan Crane, Clarisse Weischedel, Max Wardetzky"
     """
 
-    def __init__(self, shape):
-        super().__init__(shape)
-        self.solver = pp3d.MeshHeatMethodDistanceSolver(shape.vertices,shape.faces)
-        
-        
-    def dist_matrix(self):
-        """Distance between mesh vertices.
-
-        Returns
-        -------
-        dist_matrix : array-like, shape=[n_vertices, n_vertices]
-            Distance matrix.
-
-        Notes
-        -----
-        slow
-        """
-        dist_mat = np.empty((self._shape.n_vertices, self._shape.n_vertices))
-        for i in range(self._shape.n_vertices):
-            dist_mat[i] = self.solver.compute_distance(i)
-
-        return dist_mat
-
-    
-    def _dist_from_source_single(self, source_point):
-        """Distance between mesh vertices.
-
-        Parameters
-        ----------
-        source_point : array-like, shape=()
-            Index of source point.
-
-        Returns
-        -------
-        dist : array-like, shape=[n_vertices]
-            Distance.
-        target_point : array-like, shape=[n_vertices,]
-            Target index.
-        """
-        dist=self.solver.compute_distance(source_point.item())
-        
-        target_point = np.arange(self._shape.n_vertices)
-
-        return dist, target_point
-
-    def _dist_single(self, point_a, point_b):
-        """Distance between mesh vertices.
-
-        Parameters
-        ----------
-        point_a : array-like, shape=()
-            Index of source point.
-        point_b : array-like, shape=()
-            Index of target point.
-
-        Returns
-        -------
-        dist : numeric
-            Distance.
-        """
-        dist=self.solver.compute_distance(point_a)[point_b]
-        
-        return dist
+    _Registry = HeatDistanceMetricRegistry
