+++ geomfum/wrap/__init__.py
@@ -3,6 +3,7 @@ from geomfum._registry import (
     register_face_orientation_operator,
     register_face_valued_gradient,
     register_farthest_point_sampler,
+    register_heat_distance_metric,
     register_heat_kernel_signature,
     register_hierarchical_mesh,
     register_laplacian_finder,
@@ -75,4 +76,8 @@ register_mesh_plotter("pyvista", "PvMeshPlotter", requires="pyvista", as_default
 
 register_mesh_plotter(
     "polyscope", "PsMeshPlotter", requires="polyscope", as_default=False
+)
+
+register_heat_distance_metric(
+    "pp3d", "Pp3dHeatDistanceMetric", requires="potpourri3d", as_default=True
 )
